# lib/api-zod/src/generated/types/availabilitySlot.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Spectra Agency public funnel and secure admin operations
 * OpenAPI spec version: 0.1.0
 */

export interface AvailabilitySlot {
  startsAt: Date;
  endsAt: Date;
  timezone: string;
}