# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Spectra Agency public funnel and secure admin operations
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  leadId: number;
  clientName: string;
  clientEmail: string;
  startsAt: Date;
  endsAt: Date;
  timezone: string;
  status: BookingStatus;
  /** @nullable */
  meetingUrl?: string | null;
  /** @nullable */
  calendarEventId?: string | null;
  createdAt: Date;
}