# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Spectra Agency public funnel and secure admin operations
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  leadId: number;
  startsAt: Date;
  endsAt: Date;
  timezone: string;
}