# lib/api-zod/src/generated/types/bookingStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Spectra Agency public funnel and secure admin operations
 * OpenAPI spec version: 0.1.0
 */

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  requested: 'requested',
  approved: 'approved',
  cancelled: 'cancelled',
  completed: 'completed',
  no_show: 'no_show',
} as const;