# lib/api-zod/src/generated/types/bookingStatusUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Spectra Agency public funnel and secure admin operations
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface BookingStatusUpdate {
  status: BookingStatus;
}