# lib/api-zod/src/generated/types/calendarSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Spectra Agency public funnel and secure admin operations
 * OpenAPI spec version: 0.1.0
 */

export interface CalendarSummary {
  id: string;
  summary: string;
  primary: boolean;
  accessRole: string;
}