# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Spectra Agency public funnel and secure admin operations
 * OpenAPI spec version: 0.1.0
 */

export interface LeadInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 1 */
  company: string;
  email: string;
  phone?: string;
  businessType?: string;
  budget?: string;
  /** @minLength 10 */
  projectDescription: string;
  /** @minItems 1 */
  interestedServices: string[];
}