# lib/api-zod/src/generated/types/leadStatusUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Spectra Agency public funnel and secure admin operations
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStatus } from './leadStatus';

export interface LeadStatusUpdate {
  status: LeadStatus;
}