# lib/api-zod/src/generated/types/publicConfig.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Spectra Agency public funnel and secure admin operations
 * OpenAPI spec version: 0.1.0
 */
import type { PublicVideo } from './publicVideo';

export interface PublicConfig {
  video: PublicVideo | null;
  bookingDurationMinutes: number;
  timezone: string;
}