# lib/api-zod/src/generated/types/uploadRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Spectra Agency public funnel and secure admin operations
 * OpenAPI spec version: 0.1.0
 */
import type { UploadRequestContentType } from './uploadRequestContentType';

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /**
     * @minimum 1
     * @maximum 524288000
     */
  size: number;
  contentType: UploadRequestContentType;
}